Amend the CKS level-one OFI scan: add the 0.5 s response horizon, and benchmark the literature
Amendment 1 takes the exploratory grid from 25 to 30 cells by admitting h2 = 0.5 s,
after measuring that the depth-20 mid resolves to two different snapshots at both
as-of endpoints in 99.4-99.5% of cases. The h1 = 0.5 s floor is retained, on restated
grounds: e_n is zero unless the best quote changes, and best-quote changes arrive only
1.42/s and 0.77/s, so below half a second the regressor is zero in a majority of windows.

The new arm is a clean negative - no cell clears the three dependence checks and the
past-return mirror beats the future increment in all five - and it costs two headline
counts, with cells clearing all three checks falling to 1 of 30 for raw OFI and 0 of 30
for depth-scaled pressure.

Also records that the displayed level one is not the true touch on this feed: 42-48% of
executions print strictly inside the displayed best bid/ask, so the constructed object is
the imbalance of the outermost displayed band and the CKS ~65% benchmark does not apply.

deep_book_ofi.RETURN_HORIZONS_SECONDS is deliberately not widened; the frozen
X-OFI-DAT20-03 design is untouched and the horizon annotation changes are runtime-inert.

+++ src/shaurya/signals/deep_book_normal_activity.py
@@ -469,9 +469,9 @@ class Observation:
     receive_ts_ns: int
     time_bucket: str
     features: Mapping[str, float]
-    future_ticks: Mapping[int, float]
-    past_ticks: Mapping[int, float]
-    contemporaneous_ticks: Mapping[int, float]
+    future_ticks: Mapping[float, float]
+    past_ticks: Mapping[float, float]
+    contemporaneous_ticks: Mapping[float, float]
 
 
 def build_flow_features(
@@ -573,7 +573,7 @@ def response_ticks(
     anchor_ts_ns: int,
     horizons_seconds: Sequence[int] = HORIZONS_SECONDS,
     tick_size: float = FUTURES_TICK_SIZE,
-) -> tuple[dict[int, float], dict[int, float], dict[int, float]]:
+) -> tuple[dict[float, float], dict[float, float], dict[float, float]]:
     """Future, past-mirror and contemporaneous mid-price moves at one publication instant.
 
     - **future** — the move from the anchor to ``anchor + h``.  This is the thing being predicted.
@@ -595,9 +595,9 @@ def response_ticks(
     """
 
     anchor = series.as_of(anchor_ts_ns)
-    future: dict[int, float] = {}
-    past: dict[int, float] = {}
-    contemporaneous: dict[int, float] = {}
+    future: dict[float, float] = {}
+    past: dict[float, float] = {}
+    contemporaneous: dict[float, float] = {}
     if anchor is None:
         return future, past, contemporaneous
     for horizon in horizons_seconds:
